Cloud-Native E-Commerce Platform/services/order-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
kafka-python==2.0.2
pika==1.3.2
prometheus-client==0.20.0
